+++ apps/frontend/.eslintrc.cjs
@@ -91,14 +91,20 @@ module.exports = {
 	},
 	overrides: [
 		{
-			files: [
-				'./src/assets/icons.ts',
-				'./src/data/eslint-data.ts',
-				'./src/data/stylelint-data.ts',
-				'./src/data/prettier-data.ts',
-			],
+			files: ['./src/assets/icons.ts', './src/data/**/*.ts'],
 			rules: {
 				'max-lines': 'off',
+			},
+		},
+		{
+			files: [
+				'./src/components/containers/GroupCenter/NewPolicy/NewPolicy.view.tsx',
+				'./src/components/containers/GroupCenter/NewPolicy/LibrariesList/Library/Library.view.tsx',
+				'./src/components/containers/GroupCenter/GroupDetails/Policies/PoliciesList/PoliciesList.tsx',
+				'./src/i18n/en.ts',
+			],
+			rules: {
+				'max-lines': ['error', { max: 200, skipBlankLines: true, skipComments: true }],
 			},
 		},
 	],