+++ apps/frontend/src/components/containers/GroupCenter/GroupDetails/Policies/PoliciesList/interfaces/table-columns.ts
@@ -1,0 +1,6 @@
+export interface ITableColumns {
+	title: string;
+	dataIndex: string;
+	key: string;
+	width: number;
+}