+++ apps/frontend/src/components/containers/GroupCenter/NewPolicy/SideBarFilters/CategoriesButtons/index.ts
@@ -1,0 +1,3 @@
+import CategoriesButtons from './CategoriesButtons';
+
+export default CategoriesButtons;