+++ apps/frontend/src/components/containers/GroupCenter/NewPolicy/SideBarFilters/TypesButtons/index.ts
@@ -1,0 +1,3 @@
+import TypesButtons from './TypesButtons';
+
+export default TypesButtons;