+++ apps/frontend/src/data/stylelint-data.ts
@@ -1,7 +1,9 @@
+import type { ILibraryData } from '@/interfaces/libraries';
+
 import { LibraryCategory } from '../models/library-category';
 import { LibraryType } from '../models/library-type';
 
-export const stylelintData = {
+export const stylelintData: ILibraryData = {
 	name: 'Stylelint',
 	author: 'stylelint.io',
 	description: